+++ src/server/gameserver/ItemInfo.cpp
@@ -48,7 +48,7 @@ ItemInfo::~ItemInfo()
 }
 
 VolumeWidth_t ItemInfo::getVolumeWidth () const 
-	throw() 
+	throw () 
 { 
 	__BEGIN_TRY
 
@@ -58,7 +58,7 @@ VolumeWidth_t ItemInfo::getVolumeWidth () const
 }
 
 VolumeHeight_t ItemInfo::getVolumeHeight () const 
-	throw() 
+	throw () 
 { 
 	__BEGIN_TRY
 
@@ -70,7 +70,7 @@ VolumeHeight_t ItemInfo::getVolumeHeight () const
 void ItemInfo::setReqAbility(const string& req)
 	throw()
 {
-    size_t a = 0, b = 0, c = 0;
+	uint a = 0, b = 0, c = 0;
 
 	do
 	{
@@ -113,13 +113,13 @@ void ItemInfo::setDefaultOptions(const string& options)
 		return;
 	}
 
-	makeOptionList(options, m_DefaultOptions);
+	makeOptionList( options, m_DefaultOptions );
 
 	/*do
 	{
 		b = options.find_first_of(',', a);
 
-		string  optionName = trim(options.substr(a, b-a));
+		string  optionName = trim( options.substr(a, b-a) );
 
 		OptionType_t optionType;
 		
@@ -129,7 +129,7 @@ void ItemInfo::setDefaultOptions(const string& options)
 			throw Error("그런 옵션이 없단다.");
 		}
 
-		addDefaultOptions(optionType);
+		addDefaultOptions( optionType );
 
 		a = b+1;
 
@@ -141,7 +141,7 @@ void ItemInfo::addDefaultOptions(OptionType_t optionType)
 	throw()
 {
 	m_bDefaultOptions = true;
-	m_DefaultOptions.push_back(optionType);
+	m_DefaultOptions.push_back( optionType );
 }
 
 void ItemInfo::clearDefaultOptions()
