+++ src/server/gameserver/quest/VariableBuffer.cpp
@@ -23,7 +23,7 @@ const char * VariableBuffer::SpaceTab = " \t";
 // constructor
 ////////////////////////////////////////////////////////////////////////////////
 VariableBuffer::VariableBuffer (const string & buffer) 
-	throw()
+	throw ()
 {
 	m_Buffer = buffer;
 
@@ -34,7 +34,7 @@ VariableBuffer::VariableBuffer (const string & buffer)
 // destructor
 ////////////////////////////////////////////////////////////////////////////////
 VariableBuffer::~VariableBuffer () 
-	throw()
+	throw ()
 {
 	clear();
 }
@@ -43,21 +43,21 @@ void VariableBuffer::clear()
 {
 	vector<VariableInfo*>::iterator itr = m_VariableInfos.begin();
 
-	for (; itr != m_VariableInfos.end(); itr++ )
+	for ( ; itr != m_VariableInfos.end(); itr++ )
 	{
-		SAFE_DELETE(*itr);
+		SAFE_DELETE( *itr );
 	}
 
 	m_VariableInfos.clear();
 }
 
 void VariableBuffer::parse()
-	throw(Error)
+	throw (Error)
 {
 	__BEGIN_TRY
 
-	string subStr = trim(m_Buffer);
-	if (subStr.size() == 0 )
+	string subStr = trim( m_Buffer );
+	if ( subStr.size() == 0 )
 		return;
 
 	bool bEnd = false;
@@ -67,35 +67,35 @@ void VariableBuffer::parse()
 
 	do
 	{
-        size_t j = subStr.find_first_of(Separator, i);
+		uint j = subStr.find_first_of( Separator, i );
 
 		// separator 가 없다면 마지막 Variable 이다.
-		if (j == string::npos )
+		if ( j == string::npos )
 		{
 			j = subStr.size();
 			bEnd = true;
 		}
 
-		string variable = trim(subStr.substr(i, j - i ));
+		string variable = trim(subStr.substr( i, j - i ));
 
-		VariableInfo* pVariableInfo = new VariableInfo(variable); 
+		VariableInfo* pVariableInfo = new VariableInfo( variable ); 
 		
-		m_VariableInfos.push_back(pVariableInfo);
+		m_VariableInfos.push_back( pVariableInfo );
 
 		index++;
 		i = j + 1;
 	}
-	while (!bEnd);
+	while ( !bEnd );
 
 	__END_CATCH
 }
 
-VariableInfo* VariableBuffer::getVariableInfo(int index )
-	throw(OutOfBoundException, Error )
+VariableInfo* VariableBuffer::getVariableInfo( int index )
+	throw( OutOfBoundException, Error )
 {
 	__BEGIN_TRY
 
-	if (index < 0 || index >= (int)m_VariableInfos.size() )
+	if ( index < 0 || index >= (int)m_VariableInfos.size() )
 		throw OutOfBoundException("VariableBuffer::getVariableInfo() OOB!!");
 
 	return m_VariableInfos[index];
@@ -107,7 +107,7 @@ VariableInfo* VariableBuffer::getVariableInfo(int index )
 // get debug string
 ////////////////////////////////////////////////////////////////////////////////
 string VariableBuffer::toString () const 
-	throw()
+	throw ()
 {
 	__BEGIN_TRY
 
@@ -117,7 +117,7 @@ string VariableBuffer::toString () const
 
 	vector<VariableInfo*>::const_iterator itr = m_VariableInfos.begin();
 
-	for (; itr != m_VariableInfos.end(); itr++ )
+	for ( ; itr != m_VariableInfos.end(); itr++ )
 	{
 		msg << (*itr)->toString() << ",";
 	}
