+++ src/server/gameserver/quest/VariableInfo.cpp
@@ -16,7 +16,7 @@ const char VariableInfo::RParen = ')';
 const char VariableInfo::ParamSeperator = '|';
 
 
-VariableInfo::VariableInfo(const string& buffer )
+VariableInfo::VariableInfo( const string& buffer )
 {
 	m_Buffer = buffer;
 
@@ -31,19 +31,19 @@ VariableInfo::~VariableInfo()
 //
 //             i             j      k
 // 0123456789012345678901234567890123
-// EntranceFee : EntranceFee (1003 )
+// EntranceFee : EntranceFee ( 1003 )
 //
 /////////////////////////////////////////////////////////
 void VariableInfo::parse()
-	throw(Error)
+	throw (Error)
 {
 	__BEGIN_TRY
 
-    size_t i = m_Buffer.find(Seperator);
-    size_t j = m_Buffer.find(LParen);
-    size_t k = m_Buffer.find(RParen);
+	uint i = m_Buffer.find( Seperator );
+	uint j = m_Buffer.find( LParen );
+	uint k = m_Buffer.find( RParen );
 
-	if (i == string::npos
+	if ( i == string::npos
 		|| j == string::npos
 		|| k == string::npos
 		|| i > j
@@ -53,12 +53,12 @@ void VariableInfo::parse()
 		throw Error("Invalid VariableInfo String");
 	}
 
-	m_Name = trim(m_Buffer.substr(0, i ));
-	m_Keyword = trim(m_Buffer.substr(i+1, j-i-1 ));
+	m_Name = trim( m_Buffer.substr( 0, i ) );
+	m_Keyword = trim( m_Buffer.substr( i+1, j-i-1 ) );
 
-	string parameters = trim(m_Buffer.substr(j+1, k-j-1 ));
+	string parameters = trim( m_Buffer.substr( j+1, k-j-1 ) );
 
-	if (parameters.size() == 0 )
+	if ( parameters.size() == 0 )
 		return;
 
 	bool bEnd = false;
@@ -68,33 +68,33 @@ void VariableInfo::parse()
 
 	do 
 	{
-		j = parameters.find_first_of(ParamSeperator, i);
+		j = parameters.find_first_of( ParamSeperator, i );
 
 		// seperator 가 없다면 마지막 값이다.
-		if (j == string::npos )
+		if ( j == string::npos )
 		{
 			j = parameters.size();
 			bEnd = true;
 		}
 
-		string parameter = parameters.substr(i, j-i);
-		//m_Parameters.push_back(trim(parameters.substr(i, j-i ) ));
-		m_Parameters.push_back(trim(parameter));
+		string parameter = parameters.substr( i, j-i );
+		//m_Parameters.push_back( trim( parameters.substr( i, j-i ) ) );
+		m_Parameters.push_back( trim( parameter) );
 
 		index ++;
 		i = j + 1;
 	}
-	while (!bEnd);
+	while ( !bEnd );
 	
 	__END_CATCH
 }
 
-const string& VariableInfo::getParameter(int index ) const
+const string& VariableInfo::getParameter( int index ) const
 	throw(OutOfBoundException, Error)
 {
 	__BEGIN_TRY
 
-	if (index < 0 || index >= (int)m_Parameters.size() )
+	if ( index < 0 || index >= (int)m_Parameters.size() )
 		throw OutOfBoundException("VariableInfo::getParameter() OOB!!");
 
 	return m_Parameters[index];
@@ -116,7 +116,7 @@ string VariableInfo::toString() const
 
 	vector<string>::const_iterator itr = m_Parameters.begin();
 
-	for (; itr != m_Parameters.end(); itr++ )
+	for ( ; itr != m_Parameters.end(); itr++ )
 	{
 		msg << (*itr) << ",";
 	}
